crm_hcp_module/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
langgraph
langchain
langchain-core
langchain-groq
python-dotenv
